A few tweaks here and there.

+++ networks/MalariaGEN/projects/dataMerger/trunk/src/main/webapp/pages/settings/database/js/database.js
@@ -41,21 +41,30 @@ function initDeleteDatabaseFunction () {
 			
 			function () {
 				
-				$.ajax({
-					  type: "DELETE",
-					  data: '',
-					  url: "/dataMerger/data/databases",
-					  dataType: "text",
-					  success: function(data, textStatus, jqXHR) {
-							$('.ajaxResponse').html("Response: " + data);
-							$('.refreshButtonContainer').show();
-					  },
-					  error:function (jqXHR, textStatus, errorThrown){
-							$('.ajaxError').html("Error: " + errorThrown);
-			          },
-						beforeSend: function() { $('.deleting-database-indicator').show(); },
-				        complete: function() { $('.deleting-database-indicator').hide(); }
-					});
+				if (confirm("Are you sure you want to delete the entire database?")) {
+				
+					if (confirm("Are you SURE you're sure? You want to delete the entire database?")) {
+					
+						$.ajax({
+							  type: "DELETE",
+							  data: '',
+							  url: "/dataMerger/data/databases",
+							  dataType: "text",
+							  success: function(data, textStatus, jqXHR) {
+									$('.ajaxResponse').html("Response: " + data);
+									$('.refreshButtonContainer').show();
+							  },
+							  error:function (jqXHR, textStatus, errorThrown){
+									$('.ajaxError').html("Error: " + errorThrown);
+					          },
+								beforeSend: function() { $('.deleting-database-indicator').show(); },
+						        complete: function() { $('.deleting-database-indicator').hide(); }
+							});
+						}
+						
+					
+				}
+				
 			}
 			
 		);
@@ -95,23 +104,32 @@ function initDeleteDatabaseTablesFunction () {
 			
 			function () {
 				
-				$.ajax({
-					  type: "DELETE",
-					  data: '',
-					  url: "/dataMerger/scripts/data/databases/tables/delete-all",
-					  dataType: "text",
-					  success: function(data, textStatus, jqXHR) {
-							$('.ajaxResponse').html("Response: " + data);
-							$('.refreshButtonContainer').show();
-					  },
-					  error:function (jqXHR, textStatus, errorThrown){
-							$('.ajaxError').html("Error: " + errorThrown);
-			          },
-						beforeSend: function() { $('.deleting-database-tables-indicator').show(); },
-				        complete: function() { $('.deleting-database-tables-indicator').hide(); }
-					});
-			}
+				if (confirm("Are you sure you want to delete all the database tables?")) {
+					
+					if (confirm("Are you SURE you're sure? You want to delete all the database tables?")) {
+					
+						
+						$.ajax({
+							  type: "DELETE",
+							  data: '',
+							  url: "/dataMerger/scripts/data/databases/tables/delete-all",
+							  dataType: "text",
+							  success: function(data, textStatus, jqXHR) {
+									$('.ajaxResponse').html("Response: " + data);
+									$('.refreshButtonContainer').show();
+							  },
+							  error:function (jqXHR, textStatus, errorThrown){
+									$('.ajaxError').html("Error: " + errorThrown);
+					          },
+								beforeSend: function() { $('.deleting-database-tables-indicator').show(); },
+						        complete: function() { $('.deleting-database-tables-indicator').hide(); }
+							});
+					}
 			
+				}
+				
+			}
+					
 		);
 }
